% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.geogs.fxn}
\alias{fit.geogs.fxn}
\title{Fit adaptive spline models}
\usage{
fit.geogs.fxn(
  df.list,
  data.years,
  lam.mult = 2,
  lam.mult.base = 2,
  degrees.cold = 5,
  degrees.hot = 5,
  degrees.to.fit = 5
)
}
\arguments{
\item{df.list}{list containing y.df (data.frame containing the dependent variable values) and x.df (data.frame containing the independent variable values).}

\item{data.years}{vector (length: variable) of integers indicating the calendar years of the data.}

\item{lam.mult}{a multiplier indicating how much to stiffen the smooth.spline in each iteration (unless the performance criterion is achieved).}

\item{lam.mult.base}{a multiplier indicating how much to stiffen the smooth.spline on the initial fit (with 1 translating to using the lambda found via leave-one-out cross-validation).}

\item{degrees.cold}{how many degrees to extrapolate the empirical support for cold temperatures for each geography.}

\item{degrees.hot}{how many degrees to extrapolate the empirical support for hot temperatures for each geography.}

\item{degrees.to.fit}{how many degrees to use when fitting the linear regression used to extrapolate the empirical support.}
}
\value{
a list (length: 2) containing sublists holding each geography's "daily average" and "intra-day" models.
}
\description{
This function loops over geographies and calls helper functions once for each geography to learn "daily average" and "intra-day" models. It saves all results in a list.
}
