% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.and.check.data.fxn}
\alias{load.and.check.data.fxn}
\title{Load and ensure consistency of data}
\usage{
load.and.check.data.fxn(x.root, y.root, file.dates)
}
\arguments{
\item{x.root}{string indicating the structure of the filenames for the independent variable.}

\item{y.root}{string indicating the structure of the filenames for the dependent variable.}

\item{file.dates}{vector (length: variable) of integers indicating the dates of the data.}
}
\value{
a list (length: 2) containing final \code{x.df} and \code{y.df} arrays.
}
\description{
This function reads in the \code{X} and \code{Y} data and ensures consistency. It calls \code{load.data.fxn} twice (once to build the \code{x.df} and once to build the \code{y.df}) and then calls \code{ensure.compatible.arrays.fxn} to make sure \code{x.df} and \code{y.df} are consistent.
}
