% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.data.fxn}
\alias{load.data.fxn}
\title{Load data}
\usage{
load.data.fxn(type, root, file.dates, patt = "yyyy")
}
\arguments{
\item{type}{string (either "x" or "y"); used to determine the full read path of input data files.}

\item{root}{string indicating the structure of the filenames, including the extension (".csv" is expected), but with the characters indicating the date (e.g., 4-digit year) replaced by a compatible pattern as provided by "patt" (e.g., "yyyy"). filename structure is assumed consistent across files.}

\item{file.dates}{vector (length: variable) of integers indicating the dates of the data.}
}
\value{
a single data.frame (size: n.hours x n.geographies) that combines all data specified by the input args.
}
\description{
This function reads in \code{X} or \code{Y} data, one type at a time.
}
