% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.results.fxn}
\alias{plot.results.fxn}
\title{Plot "daily average" model fits for each geography}
\usage{
\method{plot}{results.fxn}(
  daily.average.list,
  data.years,
  plot.weekdays = TRUE,
  plot.weekends = TRUE
)
}
\arguments{
\item{daily.average.list}{list of lists containing fitted "daily average" models for each geography.}

\item{data.years}{numeric vector indicating the years of the data.}

\item{plot.weekdays}{boolean indicating whether to plot the "weekday" models.}

\item{plot.weekends}{boolean indicating whether to plot the "weekend" models.}
}
\description{
This function plots the "weekday" and "weekend" "daily average" spline predictions against the empirical values to allow the user to assess goodness of fit. Automatically determines reasonable plot dimensions and plots each geography on consistent x-axis limits.
}
