% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fxn}
\alias{predict.fxn}
\title{Predict using adaptive spline models}
\usage{
\method{predict}{fxn}(x.df, first.date, from.csv = TRUE, fitted.models = NULL)
}
\arguments{
\item{x.df}{a data.frame (n.hours x n.geogs) of forecast independent variable values to serve as inputs for prediction.}

\item{first.date}{the date of the first day to be predicted, in "YYYY-MM-DD" structure.}

\item{from.csv}{boolean indicating whether to read the fitted models from CSV files (TRUE) or whether they are in working memory already (FALSE).}

\item{fitted.models}{optional arg (defaulting to NULL) containing the fitted models; required when "from.csv" is FALSE.}
}
\value{
a data.frame of hourly predictions for the intersection of geogs included in "x.df" and the fitted models.
}
\description{
This function makes hourly predictions from fitted models using new "x.df" data.
}
