% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write.preds.fxn}
\alias{write.preds.fxn}
\title{Write model predictions}
\usage{
write.preds.fxn(preds.df, first.date)
}
\arguments{
\item{preds.df}{a data.frame (n.hours x n.geogs) of predicted output variable values.}

\item{first.date}{string formatted as "YYYYMMDD" indicating the first date for which predictions were generated.}
}
\description{
This function writes out hourly predictions from fitted models using new "x.df" data.
}
