% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write.results.fxn}
\alias{write.results.fxn}
\title{Load and ensure consistency of data}
\usage{
write.results.fxn(results.list, write.daily.avg = TRUE, write.intra.day = TRUE)
}
\arguments{
\item{results.list}{list of lists containing fitted models for each geography.}

\item{write.daily.avg}{boolean indicating whether to write out the fitted "daily average" models.}

\item{write.intra.day}{boolean indicating whether to write out the fitted "intra-day" models.}
}
\description{
This function reads in the \code{X} and \code{Y} data and ensures consistency. It calls \code{load.data.fxn} twice (once to build the \code{x.df} and once to build the \code{y.df}) and then calls \code{ensure.compatible.arrays.fxn} to make sure \code{x.df} and \code{y.df} are consistent.
}
